yaffs2: Limited yaffs2 rt_many because of O(n) in-RAM linked-list
I was confused why yaffs2 was taking so long to run the
bench_p26_rt_many, in theory it should be the fastest filesystem even at
scale. But it turns out the in-RAM linked-list that backs directories
became a bottleneck! At O(n) the whole benchmark runs O(n^2).

This isn't what we're benchmarking, and in theory may not matter when IO
costs are involved, but it's a practical problem for our simulated
benchmarks.

Note this limit can be bypassed with bench.py's -a/--all flag.