Added flash clobbering to try to level playing field
This explains some confusing results in our benchmarks on large disk
sizes

As it was, spiffs and yaffs2 both fully erase the disk during flash as a
part of format. This isn't bad design, littlefs will need to erase
eventually, and flash often has mass erase operations. The problem is
that we weren't including any format operations in our benchmarks. This
skews results in spiffs/yaffs2 favor, but only if the disk is large
enough (or benchmark is small enough) to avoid cycling the disk and
triggering gc.

To avoid this, I added bench_helpers_clobber, which clobbers the disk
after format with the intention of forcing spiffs and yaffs2 to do gc.
In theory this isn't unfair to spiffs or yaffs2 as they would need to do
this work eventually.

Currently the benchmarks assume an IO bottlenecked application in which
background gc work is not possible. If we want to consider gc, we should
probably include littlefs's gc, but I think it's low priority at the
moment.

Another alternative would be to run the benchmarks long enough for this
preerased storage to have minimal effect, but that may make benchmarking
difficult.