Reverted yaffs2's warmup workaround
It turns out the shortened writes were a symptom of an ENOSPC error,
that got twisted into EINVAL by yaffs internals somehow?

Not sure at all what's going on, but I'm started to suspect a yaffs bug
related to >=1MiB block sizes. It may just be an issue with gc tuning,
but other things seem to also be breaking...

Just reverting to the assert for now, with a comment. 1MiB block sizes
may be out of reach for our current yaffs2 benchmarks.