Reduced spiffs/yaffs2 cache lines 3 -> 2
This doesn't really have a noticeable effect, other than decreasing RAM
usage. This is probably because we're measuring relatively simple write
patterns and aren't very heavy on reads.

This is more fair to spiffs, as it also pays for a 2 page "work buffer",
similar to littlefs's read/prog caches.

More fair would probably be a cache line of 1, but this caused spiffs to
spazz out, so cache line = 2 is probably the lower limit.